Creating app specialties and making new migrations
