+++ relearn/settings.py
@@ -29,7 +29,8 @@ INSTALLED_APPS = [
     'django.contrib.staticfiles',
     'users.apps.UsersConfig',
     'stackQA.apps.StackqaConfig',
-    'likes.apps.LikesConfig'
+    'likes.apps.LikesConfig',
+    'specialties.apps.SpecialtiesConfig',
 ]
 
 MIDDLEWARE = [
