+++ specialties/views.py
@@ -1,0 +1,6 @@
+from django.http import HttpResponse
+from django.shortcuts import render
+
+# Create your views here.
+def home(request):
+    return HttpResponse('Specialty home')
